+++ Source/CesiumRuntime/Private/GeoTransforms.cpp
@@ -7,6 +7,8 @@
 #include "CesiumTransforms.h"
 #include "VecMath.h"
 #include <glm/gtc/matrix_inverse.hpp>
+
+#define GLM_ENABLE_EXPERIMENTAL
 #include <glm/gtx/quaternion.hpp>
 
 using namespace CesiumGeospatial;