+++ extern/vcpkg-overlays/triplets/x64-osx-unreal.cmake
@@ -1,0 +1,3 @@
+include("${CMAKE_CURRENT_LIST_DIR}/shared/osx-common.cmake")
+set(VCPKG_TARGET_ARCHITECTURE x64)
+set(VCPKG_OSX_ARCHITECTURES x86_64)