+++ CHANGES.md
@@ -2,9 +2,14 @@
 
 ### ? - ?
 
+##### Additions :tada:
+
+- Added universal (Intel and Apple Silicon) binaries for Unreal Engine 5.2. Unreal Engine 5.3 and 5.4 already had universal binaries.
+
 ##### Fixes :wrench:
 
 - Fixed a bug introduced in the previous release that caused instanced tilesets to render incorrectly.
+- Debug sections are no longer compressed on Linux and Android, improving compatibility.
 
 ### v2.7.1 - 2024-08-01